+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.21.21
 
 - `PanaTags` fields provide a single place for most of the tags that a package can get.
+- Experimental options to analyze and tag compatibility with a future SDK.
 
 ## 0.21.20